Remove unnecessary cast
git-svn-id: https://svn.apache.org/repos/asf/jakarta/jmeter/trunk@325300 13f79535-47bb-0310-9956-ffa450edef68

Former-commit-id: 3d42b9b430d0403b1092a930190be888f2ce256a

+++ src/jorphan/org/apache/jorphan/gui/ComponentUtil.java
@@ -90,10 +90,10 @@ public final class ComponentUtil
         Component toBeCentered)
     {
         toBeCentered.setLocation(
-            (int) parent.getX()
-                + (int) ((parent.getWidth() - toBeCentered.getWidth()) / 2),
-            (int) parent.getY()
-                + (int) ((parent.getHeight() - toBeCentered.getHeight()) / 2));
+            parent.getX()
+                + (parent.getWidth() - toBeCentered.getWidth()) / 2,
+            parent.getY()
+                + (parent.getHeight() - toBeCentered.getHeight()) / 2);
 
         toBeCentered.validate();
         toBeCentered.repaint();
